fix: export used prop types
