--- src/components/Checkbox/CheckboxListInput.d.ts
@@ -1,10 +1,0 @@
-import { Component, ReactNode } from 'react';
-
-interface CheckboxListInputProps {
-  children?: ReactNode;
-  onChange?: (value: string[]) => void;
-  value?: string[];
-}
-
-declare class CheckboxListInput extends Component<CheckboxListInputProps, {}> {}
-export default CheckboxListInput;
