--- src/components/Input/FileInput.d.ts
@@ -1,8 +1,0 @@
-import * as React from 'react';
-
-interface FileInputProps
-  extends Omit<React.InputHTMLAttributes<HTMLInputElement>, 'onChange' | 'type'> {
-  onChange?: (files: FileList) => void;
-}
-declare class FileInput extends React.Component<FileInputProps, {}> {}
-export default FileInput;
